refactor(help): move help commands to help task and create a util to house helper functions.
