+++ lib/cli.js
@@ -5,25 +5,23 @@ var IonicStore = require('./utils/store');
 var IonicConfig = new IonicStore('ionic.config');
 var Info = IonicAppLib.info;
 var IonicProject = IonicAppLib.project;
-var Ionitron = require('./utils/ionitron');
 var optimist = require('optimist');
 var path = require('path');
 var fs = require('fs');
 var settings = require('../package.json');
-var Tasks = require('./tasks/cliTasks');
+var Tasks = require('./config/commands').allCommands;
 var Utils = IonicAppLib.utils;
 var Logging = IonicAppLib.logging;
 var log = Logging.logger;
 var Q = require('q');
-var semver = require('semver');
-var gutil = require('gulp-util');
-var prettyTime = require('pretty-hrtime');
+var helpUtil = require('./utils/help');
 
 var TASKS = Cli.Tasks = Tasks;
 
 Cli.IONIC_DASH = 'https://apps.ionic.io';
 Cli.IONIC_API = '/api/v1/';
 Cli.PRIVATE_PATH = '.ionic';
+Cli.VERSION = settings.version;
 
 // This is where the logger for the app-lib can be configured (or replaced).
 log.level = 'info';
@@ -79,6 +77,7 @@ Cli.run = function run(processArgv) {
      *       and this should just map the option to a command
      */
     if (argv.ionitron) {
+      var Ionitron = require('./utils/ionitron');
       var lang = argv.es ? 'es' : 'en';
       return Q.fcall(Ionitron.print, lang);
     }
@@ -99,7 +98,7 @@ Cli.run = function run(processArgv) {
      *       $ ionic start --help
      */
     if (argv.help || argv.h) {
-      return Q.fcall(Cli.printHelpLines);
+      return Cli.printHelpLines();
     }
 
     /*
@@ -119,27 +118,24 @@ Cli.run = function run(processArgv) {
     IonicStats.t();
 
     var taskName = argv._[0];
-    var taskSetting = Cli.getTaskByName(taskName);
+    var task = Cli.getTaskSettingsByName(taskName);
 
     // If no task is found then just provide the list of tasks available
-    if (!taskSetting) {
-      return Q.fcall(Cli.printAvailableTasks, taskName);
+    if (!task) {
+      return helpUtil.printTaskListShortUsage(Cli.getAllTaskSettings(), taskName, Cli.VERSION);
     }
 
-    log.debug('Task setting:', taskSetting);
+    log.debug('Task:', task);
 
-    var booleanOptions = Cli.getListOfBooleanOptions(taskSetting.options);
+    var booleanOptions = Cli.getListOfBooleanOptions(task.options);
 
     // Remove all possible task boolean options from the args
     argv = optimist(processArgv.slice(2))
       .boolean(booleanOptions)
       .argv;
 
-    // Get the task 'class'
-    var taskInstance = Cli.lookupTask(taskSetting.module);
-
     // this should be tracked by each task
-    if (!taskSetting.disableChangePwd) {
+    if (!task.disableChangePwd) {
       var root = Utils.cdIonicRoot();
       var project = IonicProject.load(root);
       argv.v2 = project && project.get('v2');
@@ -148,12 +144,12 @@ Cli.run = function run(processArgv) {
       if (fs.existsSync('node_modules')) {
 
         // run with gulp hooks
-        return runWithGulp(argv, taskInstance);
+        return runWithGulp(argv, task);
       } else if (argv.v2) {
         log.warn('WARN: No node_modules directory found, do you need to run npm install?'.yellow);
       }
     }
-    return Q.fcall(taskInstance.run, Cli, argv);
+    return Q.fcall(task.run, Cli, argv);
 
   } catch (ex) {
     return Utils.fail(ex);
@@ -346,6 +342,9 @@ function installGulpfilePackages() {
 
 
 function logEvents(gulpInst, finalTaskNames) {
+  var prettyTime = require('pretty-hrtime');
+  var gutil = require('gulp-util');
+
   gulpInst.on('task_start', function(e) {
 
     // TODO: batch these
@@ -504,209 +503,35 @@ Cli.checkLatestVersion = function checkLatestVersion() {
 
 
 /**
- * @method getTaskByName
+ * @method getTaskSettingsByName
  * @param {String} taskName task name to look for
  * @return {Object} Returns the task settings object that matches
  */
-Cli.getTaskByName = function getTaskByName(taskName) {
-  return TASKS.find(function(task) {
-    return (task.name === taskName) || (Array.isArray(task.alt) && task.alt.indexOf(taskName) !== -1);
+Cli.getTaskSettingsByName = function getTaskSettingsByName(taskName) {
+  var task;
+
+  Object.keys(TASKS).every(function(listName) {
+    if (listName === taskName) {
+      task = require(TASKS[listName]);
+      return false;
+    }
+    return true;
+  });
+
+  return task;
+};
+
+
+/**
+ * @method getAllTaskSettings
+ * @return {Array} Returns an array of task settings objects
+ */
+Cli.getAllTaskSettings = function getAllTaskSettings() {
+  return Object.keys(TASKS).map(function(listName) {
+    return require(TASKS[listName]);
   });
 };
 
-Cli.printIonic = function printIonic() {
-  return [
-    '  _             _          ',
-    ' (_)           (_)         ',
-    '  _  ___  _ __  _  ___     ',
-    ' | |/ _ \\| \'_ \\| |/ __| ',
-    ' | | (_) | | | | | (__     ',
-    ' |_|\\___/|_| |_|_|\\___|  CLI v' + settings.version
-  ];
-};
-
-function printTemplate(lineArray) {
-  console.log(lineArray.reduce(function(all, line) {
-    all += (line !== null) ? line + '\n' : '';
-    return all;
-  }, ''));
-}
-
-
-/**
- * @method printAvailableTasks
- * @param {String} taskName task name supplied to cli
- * @return {Null} no return value
- */
-Cli.printAvailableTasks = function printAvailableTasks(taskName) {
-  var taskLines = TASKS
-    .filter(function(task) {
-      return task.summary;
-    })
-    .map(function(task) {
-      var name = '   ' + task.name + '  ';
-      var dots = '';
-      while ((name + dots).length < 20) {
-        dots += '.';
-      }
-      return name.green.bold + dots.grey + '  ' + task.summary.bold;
-    });
-
-  var lines = [].concat(
-    Cli.printIonic(),
-    [
-      '',
-      'Usage: ionic task args',
-      '',
-      '',
-      '=======================',
-      (taskName ?
-        (taskName + ' is not a valid task\n\n').bold.red
-        : null),
-      'Available tasks: '.bold,
-      '(use --help or -h for more info)',
-      '',
-      ''
-    ],
-    taskLines
-  );
-
-  printTemplate(lines);
-  Cli.processExit(1);
-};
-
-
-/**
- * @method printHelpLines
- * @return {Null} no return value
- */
-Cli.printHelpLines = function printHelpLines() {
-  var lines = [].concat(
-    Cli.printIonic(),
-    [
-      '',
-      '=======================',
-      ''
-    ]
-  );
-  printTemplate(lines);
-
-  TASKS
-    .filter(function(task) {
-      return task.summary;
-    }).forEach(function(task) {
-      return Cli.printUsage(task);
-    });
-
-  Cli.processExit(1);
-};
-
-
-Cli.printUsage = function printUsage(d) {
-  function w(s) {
-    process.stdout.write(s);
-  }
-
-  w('\n');
-
-  var rightColumn = 45;
-  var dots = '';
-  var indent = '';
-  var x;
-  var arg;
-
-  var taskArgs = d.title;
-
-  for (arg in d.args) {
-    if ({}.hasOwnProperty.call(d.args, arg)) {
-      taskArgs += ' ' + arg;
-    }
-  }
-
-  w(taskArgs.green.bold);
-
-  while ((taskArgs + dots).length < rightColumn + 1) {
-    dots += '.';
-  }
-
-  w(' ' + dots.grey + '  ');
-
-  if (d.summary) {
-    w(d.summary.bold);
-  }
-
-  for (arg in d.args) {
-    if (!d.args[arg]) continue;
-
-    indent = '';
-    w('\n');
-    while (indent.length < rightColumn) {
-      indent += ' ';
-    }
-    w((indent + '    ' + arg + ' ').bold);
-
-    var argDescs = d.args[arg].split('\n');
-    var argIndent = indent + '    ';
-
-    for (x = 0; x < arg.length + 1; x += 1) {
-      argIndent += ' ';
-    }
-
-    for (x = 0; x < argDescs.length; x += 1) {
-      if (x === 0) {
-        w(argDescs[x].bold);
-      } else {
-        w('\n' + argIndent + argDescs[x].bold);
-      }
-    }
-  }
-
-  indent = '';
-  while (indent.length < d.name.length + 1) {
-    indent += ' ';
-  }
-
-  var optIndent = indent;
-  while (optIndent.length < rightColumn + 4) {
-    optIndent += ' ';
-  }
-
-  for (var opt in d.options) {
-    if ({}.hasOwnProperty.call(d.options, opt)) {
-      w('\n');
-      dots = '';
-
-      var optLine = indent + '[' + opt + ']  ';
-
-      w(optLine.yellow.bold);
-
-      if (d.options[opt]) {
-        while ((dots.length + optLine.length - 2) < rightColumn) {
-          dots += '.';
-        }
-        w(dots.grey + '  ');
-
-        var taskOpt = d.options[opt];
-        var optDescs;
-
-        if (typeof taskOpt == 'string') {
-          optDescs = taskOpt.split('\n');
-        } else {
-          optDescs = taskOpt.title.split('\n');
-        }
-        for (x = 0; x < optDescs.length; x += 1) {
-          if (x === 0) {
-            w(optDescs[x].bold);
-          } else {
-            w('\n' + optIndent + optDescs[x].bold);
-          }
-        }
-      }
-    }
-
-    w('\n');
-  }
-};
 
 Cli.processExit = function processExit(code) {
   if (Cli.cliNews && Cli.cliNews.promise) {
@@ -721,9 +546,11 @@ Cli.processExit = function processExit(code) {
   }
 };
 
+
 Cli.version = function version() {
   log.info(settings.version + '\n');
 };
+
 
 Cli.printNewsUpdates = function printNewsUpdates(skipNewsCheck) {
   if (typeof skipNewsCheck == 'undefined') {
@@ -834,6 +661,7 @@ Cli.getContentSrc = function getContentSrc() {
 };
 
 Cli.doRuntimeCheck = function doRuntimeCheck(version) {
+  var semver = require('semver');
   var lastVersionChecked = IonicConfig.get('lastVersionChecked');
   var versionHasBeenChecked;
