+++ lib/config/commands.js
@@ -1,0 +1,80 @@
+'use strict';
+
+var extend = Object.assign || require('util')._extend; // eslint-disable-line no-underscore-dangle
+var path = require('path');
+
+var COMMAND_DIRECTORY = 'ionic';
+
+/**
+ * List of shorthands that are available for each command
+ */
+var shorthands = {
+  g: 'generate',
+  ls: 'list',
+  rm: 'remove',
+  up: 'upload'
+};
+
+/**
+ * List of affordances that are available for each command
+ * Sometimes we will add misspellings here or if we change a command name
+ */
+var affordances = {
+  platforms: 'platform',
+  plugins: 'plugin',
+  address: 'serve'
+};
+
+/**
+ * List of commands that are available from ionic cli
+ * Each command as a 1-to-1 mapping to a module with the same name
+ */
+var cmdList = [
+  'add',
+  'browser',
+  'build',
+  'compile',
+  'docs',
+  'emulate',
+  'generate',
+  'help',
+  'hooks',
+  'info',
+  'io',
+  'lib',
+  'link',
+  'list',
+  'login',
+  'package',
+  'platform',
+  'plugin',
+  'prepare',
+  'push',
+  'remove',
+  'resources',
+  'resources',
+  'run',
+  'security',
+  'serve',
+  'service',
+  'setup',
+  'share',
+  'start',
+  'state',
+  'upload'
+];
+
+var cmdMap = cmdList.reduce(function(completeObj, commandName) {
+  completeObj[commandName] = commandName;
+  return completeObj;
+}, {});
+
+var allCommands = extend(extend(extend({}, shorthands), affordances), cmdMap);
+allCommands = Object.keys(allCommands).reduce(function(allCmds, commandName) {
+  allCmds[commandName] = './' + path.join(COMMAND_DIRECTORY, allCommands[commandName]);
+  return allCmds;
+}, {});
+
+module.exports = {
+  allCommands: allCommands
+};
