+++ lib/ionic/config.js
