+++ lib/ionic/help.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var _ = require('underscore');
+var helpUtil = require('../utils/help');
 
 var settings = {
   title: 'help',
@@ -13,24 +14,21 @@ var settings = {
 };
 
 function run(ionic, argv) {
-  var command = argv._[1] ? argv._[1] : '';
+  var taskName = argv._[1] ? argv._[1] : '';
 
-  var task = ionic.getTaskWithName(command);
-
-  if (command === '') {
-    ionic.printHelpLines();
-    return;
-  } else if (!task) {
-    console.log('Command not found.'.red.bold);
-    return;
+  // If no task name is provided then print all help available
+  if (taskName === '') {
+    var taskList = ionic.getAllTaskSettings();
+    return helpUtil.printTaskListUsage(taskList, ionic.VERSION);
   }
 
-  ionic.printIonic();
-  process.stderr.write('\n=======================\n');
+  // If a task name is provided and is valid then print Usage
+  var task = ionic.getTaskSettingsByName(taskName);
+  if (task) {
+    return helpUtil.printTaskUsage(task, ionic.VERSION);
+  }
 
-  ionic.printUsage(task);
-
-  process.stderr.write('\n\n');
+  return console.log('Command not found.'.red.bold);
 }
 
 module.exports = _.extend(settings, {
