+++ lib/ionic/share.js
@@ -1,13 +1,14 @@
 'use strict';
 
 var _ = require('underscore');
-var IonicProject = require('./project');
 var LoginTask = require('./login');
 var IonicAppLib = require('ionic-app-lib');
+var IonicProject = IonicAppLib.project;
 var Share = IonicAppLib.share;
 var log = IonicAppLib.logging.logger;
 var Login = IonicAppLib.login;
 var fail = IonicAppLib.utils.fail;
+
 
 var settings = {
   title: 'share',
