+++ lib/utils/cordova.js
@@ -3,7 +3,7 @@
 var Q = require('q');
 var _ = require('underscore');
 var path = require('path');
-var colors = require('colors/safe');
+var colors = require('colors');
 var fs = require('fs');
 var IonicAppLib = require('ionic-app-lib');
 var Serve = IonicAppLib.serve;
