+++ lib/utils/ionitron.js
@@ -1,5 +1,6 @@
-require('colors');
+'use strict';
 
+var colors = require('colors');
 var path = require('path');
 var fs = require('fs');
 
@@ -31,11 +32,11 @@ var ionitronStatementsEs = [
   'Bif\u00Farcate! Oh, Lo siento, rama equivocada. '
 ];
 
-var ionictronAsciiFile = 'ionitron.txt';
+var ASCII_ART_FILE = 'ionitron.txt';
 
-module.exports.print = function print(lang) {
+function print(lang) {
 
-  var ionitronPath = path.join(__dirname, 'assets', ionictronAsciiFile);
+  var ionitronPath = path.join(__dirname, ASCII_ART_FILE);
   var contents = fs.readFileSync(ionitronPath, 'utf8');
   var messageContent;
 
@@ -66,7 +67,11 @@ module.exports.print = function print(lang) {
     }
   }
   contents = contents.replace(replaceString, messageContent);
-  console.log(contents.cyan);
+  console.log(colors.cyan(contents));
 
   return;
+}
+
+module.exports = {
+  print: print
 };
