+++ spec/cli.spec.js
@@ -1,8 +1,8 @@
 var IonicAppLib = require('ionic-app-lib');
-var Ionitron = require('../lib/ionic/ionitron');
+var Ionitron = require('../lib/utils/ionitron');
 var IonicCli = require('../lib/cli');
 var Q = require('q');
-var IonicStats = require('../lib/ionic/stats').IonicStats;
+var IonicStats = require('../lib/utils/stats').IonicStats;
 var Info = IonicAppLib.info;
 var Utils = IonicAppLib.utils;
 var Project = IonicAppLib.project;
